+++ NEWS.md
@@ -1,3 +1,5 @@
+# depigner 0.2.0
+
 * added `paired_test_categorical()` function to implement paired test
   for categorical variable accordingly to the number of groups, in a 
   suitable way to be used in the **Hmisc** `summary()` with